requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
python-dateutil==2.8.2
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
